# Backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/user": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/edituser": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditUserModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/confirmBecomeArtist": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BecomeArtistModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/cancleBecomeArtist": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BecomeArtistModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/requests": {
      "get": {
        "tags": [
          "Admin"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RequestResponse"
              }
            }
          }
        }
      }
    },
    "/Album/artist/{artistId}": {
      "get": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "artistId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Album/album/{albumId}": {
      "get": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "albumId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Album/photo/{albumId}": {
      "get": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "albumId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Album/{albumId}": {
      "delete": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "albumId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Album"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "albumId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Album/song/{albumId}/{songId}": {
      "delete": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "albumId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "songId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Album/song/{songId}": {
      "get": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "songId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Audio/{id}": {
      "get": {
        "tags": [
          "Audio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Authorization/register": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegistrate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Authorization/login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Authorization/validateToken": {
      "get": {
        "tags": [
          "Authorization"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/{userId}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Playlist"
              }
            }
          }
        }
      }
    },
    "/Playlist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreatePlaylistRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/{playlistId}/AddSong": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlistId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/uploadPhoto": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "photoFile",
            "type": "string",
            "format": "byte"
          },
          {
            "in": "formData",
            "name": "playlistId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/photo": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "playlistId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "format": "byte",
              "type": "string"
            }
          }
        }
      }
    },
    "/Playlist/Playlist/{playlistId}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlistId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/{playlistId}/RemoveSong/{songId}": {
      "delete": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlistId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "songId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/{playlistId}": {
      "delete": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlistId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Playlist/editplaylist": {
      "put": {
        "tags": [
          "Playlist"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "playlistId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Search/all": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "data",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Song/{id}": {
      "get": {
        "tags": [
          "Song"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Song": {
      "get": {
        "tags": [
          "Song"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { }
            }
          }
        }
      }
    },
    "/Song/user/{userId}": {
      "get": {
        "tags": [
          "Song"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { }
            }
          }
        }
      }
    },
    "/Song/photo/{id}": {
      "get": {
        "tags": [
          "Song"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Song/photo/songId/{id}": {
      "get": {
        "tags": [
          "Song"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Upload/upload": {
      "post": {
        "tags": [
          "Upload"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "ArtistId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "AlbumTitle",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "AudioFiles",
            "type": "array",
            "items": {
              "format": "binary",
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "PhotoFile",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "SongTitles",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "GenreIds",
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/username/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/email/{email}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/id/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/uploadAvatar": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "format": "binary",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/avatar/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/edituser": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/becomeArtist": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BecomeArtistModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Album": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "dateShared": {
          "format": "date",
          "type": "string"
        },
        "albumSongs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlbumSongs"
          }
        },
        "albumGenres": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlbumGenre"
          }
        }
      },
      "additionalProperties": false
    },
    "AlbumGenre": {
      "type": "object",
      "properties": {
        "albumId": {
          "format": "int32",
          "type": "integer"
        },
        "album": {
          "$ref": "#/definitions/Album"
        },
        "genreId": {
          "format": "int32",
          "type": "integer"
        },
        "genre": {
          "$ref": "#/definitions/Genre"
        }
      },
      "additionalProperties": false
    },
    "AlbumSongs": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "albumId": {
          "format": "int32",
          "type": "integer"
        },
        "songId": {
          "format": "int32",
          "type": "integer"
        },
        "album": {
          "$ref": "#/definitions/Album"
        },
        "song": {
          "$ref": "#/definitions/Song"
        }
      },
      "additionalProperties": false
    },
    "Audio": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "song": {
          "format": "int32",
          "type": "integer"
        },
        "duration": {
          "format": "float",
          "type": "number"
        },
        "filePath": {
          "type": "string"
        },
        "songNavigation": {
          "$ref": "#/definitions/Song"
        }
      },
      "additionalProperties": false
    },
    "BecomeArtistModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreatePlaylistRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "EditUserModel": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Genre": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "albumGenres": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlbumGenre"
          }
        }
      },
      "additionalProperties": false
    },
    "Playlist": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "playlistSongs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PlaylistSong"
          }
        }
      },
      "additionalProperties": false
    },
    "PlaylistSong": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "playlistId": {
          "format": "int32",
          "type": "integer"
        },
        "songId": {
          "format": "int32",
          "type": "integer"
        },
        "playlist": {
          "$ref": "#/definitions/Playlist"
        },
        "song": {
          "$ref": "#/definitions/Song"
        }
      },
      "additionalProperties": false
    },
    "RequestResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "userUsername": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Song": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "artist": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "albumTitle": {
          "type": "string"
        },
        "dateShared": {
          "format": "date",
          "type": "string"
        },
        "audios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Audio"
          }
        },
        "songGenres": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlbumGenre"
          }
        }
      },
      "additionalProperties": false
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserRegistrate": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}